Added LCD representation for digits 0-9
The LCD representation of digits from 0-9 can be displayed on the
console using the characters - space, hyphen and pipe characters.
Each digit is a multi-line representation of these characters. A
digit is represented as an array of lines, and each one of these
lines are made up of the above characters.